server: fix wrongly limited return data
A server can have multiple disks, thus multiple IO stats.
By limiting to the number of seconds we limit the max return as if
there was only 1 disk in the system. (1 IO per seconds where we can
have X IO per seconds with X equals the number of disks in the
system). Same for the IoNet (with interfaces).