alerts: use config files for alerts instead of the DB
Config file for the alerts will be stored in ALERTS_PATH and be
in the JSON format. Documentation will follow to help you build
those files.

In the future a "marketplace" will be added for everyone to share
their own alerts and thus forming a community to help monitor
servers with efficience.