alerts's API to server, multiple sources for alerts
First this change moves the Alerts's API endpoint to the SP-Server.
Now all endpoints are in the SP-Server and we'll only need to protect
the route from there.

Second, we now also support multiple source for the Alerts.
"files" and "database" with the latter being added for SAAS.
The doc will need to be updated accordingly, but that'll be
for latter.